% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(filename, ana_psr, clado_is_psr, clado_main_psr,
  crown_age, dme_params_filename, ext_is_psr, ext_main_psr, mig_to_is_psr,
  carry_cap_is, carry_cap_main, rng_seed, init_n_main_clades,
  init_n_main_sps)
}
\arguments{
\item{filename}{name of the file}

\item{ana_psr}{anagenesis, per-species rate}

\item{clado_is_psr}{cladogenesis on island, per-species rate}

\item{clado_main_psr}{cladogenesis on mainland, per-species rate}

\item{crown_age}{crown age, time units}

\item{dme_params_filename}{name of the file with the daisieme parameters}

\item{ext_is_psr}{extinction rate on island, per-species rate}

\item{ext_main_psr}{extinction rate on mainland, per-species rate}

\item{mig_to_is_psr}{migratrion to island, per-species rate}

\item{carry_cap_is}{carrying capacity of the island, number of species}

\item{carry_cap_main}{carrying capacity of the mainland, number of species}

\item{rng_seed}{RNG seed}

\item{init_n_main_clades}{initial number of mainland clades}

\item{init_n_main_sps}{initial number of mainland species}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
  \code{@noRd}. This is not done, as this will disallow all
  functions to find the documentation parameters
}
\author{
Richèl J.C. Bilderbeek
}
