% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dme_create_params_file.R
\name{dme_create_params_file}
\alias{dme_create_params_file}
\title{Creates a daisieme parameter file}
\usage{
dme_create_params_file(filename, ana_psr)
}
\arguments{
\item{filename}{name of the file}

\item{ana_psr}{anagenesis per-species rate}
}
\description{
Creates a daisieme parameter file
}
