% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dme_create_params_file.R
\name{dme_create_test_params_file}
\alias{dme_create_test_params_file}
\title{Creates a testing daisieme parameter file}
\usage{
dme_create_test_params_file(
  filename = tempfile(),
  ana_psr = 0.1,
  clado_is_psr = 0.1,
  clado_main_psr = 0.1,
  ext_is_psr = 0.01,
  ext_main_psr = 0.01,
  mig_to_is_psr = 0.1,
  carry_cap_is = 5,
  carry_cap_main = 20,
  rng_seed = 42,
  init_n_main_clades = 1,
  init_n_main_sps = 10,
  crown_age = 15
)
}
\arguments{
\item{filename}{name of the file}

\item{ana_psr}{anagenesis, per-species rate}

\item{clado_is_psr}{cladogenesis on island, per-species rate}

\item{clado_main_psr}{cladogenesis on mainland, per-species rate}

\item{ext_is_psr}{extinction rate on island, per-species rate}

\item{ext_main_psr}{extinction rate on mainland, per-species rate}

\item{mig_to_is_psr}{migratrion to island, per-species rate}

\item{carry_cap_is}{carrying capacity of the island, number of species}

\item{carry_cap_main}{carrying capacity of the mainland, number of species}

\item{rng_seed}{RNG seed}

\item{init_n_main_clades}{initial number of mainland clades}

\item{init_n_main_sps}{initial number of mainland species}

\item{crown_age}{crown age, time units}
}
\description{
Creates a testing daisieme parameter file
}
