% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dme_create_params_file_cpp}
\alias{dme_create_params_file_cpp}
\title{Create a parameter file.}
\usage{
dme_create_params_file_cpp(filename, ana_psr)
}
\arguments{
\item{filename}{name of the file}

\item{ana_psr}{anagenesis per-species rate}
}
\value{
nothing
}
\description{
Create a parameter file.
}
\author{
Richel J.C. Bilderbeek
}
